+++ client/src/components/Layout.tsx
@@ -1,7 +1,5 @@
 import { FC, PropsWithChildren } from 'react';
 import Sidebar from './Sidebar';
-import SearchAndAddFriend from './SearchAndAddFriend';
-import SuggestionsAccList from './SuggestionAccList';
 import { mergeClassNames } from '../utils';
 
 const Layout: FC<
@@ -14,22 +12,20 @@ const Layout: FC<
     <div className="flex w-full h-screen overflow-hidden">
       <Sidebar />
       <div className="flex-1 h-screen overflow-hidden scrollbar-hide p-10">
-        <div className="flex gap-10 mx-auto w-screen max-w-[60vw] h-screen overflow-hidden pb-20">
+        <div className="flex justify-center gap-10 mx-auto w-screen max-w-[60vw] h-screen overflow-hidden pb-20">
           <main
             className={mergeClassNames(
-              'max-w-[60%] flex-auto overflow-y-auto overflow-x-clip pr-4',
+              'max-w-[70%] flex-auto overflow-y-auto overflow-x-clip pr-4',
               mainClassName,
             )}
           >
             {children}
           </main>
-          <aside className="w-[40%] max-w-[300px] h-screen">
-            {stickyRightSideCmp !== undefined ? (
-              stickyRightSideCmp
-            ) : (
-              <DefaultStickyRightSide />
-            )}
-          </aside>
+          {stickyRightSideCmp && (
+            <aside className="w-[30%] max-w-[300px] h-screen">
+              {stickyRightSideCmp}
+            </aside>
+          )}
         </div>
       </div>
     </div>
@@ -37,12 +33,3 @@ const Layout: FC<
 };
 
 export default Layout;
-
-const DefaultStickyRightSide: FC = () => {
-  return (
-    <div className="space-y-4">
-      <SearchAndAddFriend />
-      <SuggestionsAccList />
-    </div>
-  );
-};
