+++ client/src/pages/groups/create_group/FormPanel.tsx
@@ -75,7 +75,7 @@ const GroupFormPanel = () => {
         const data = await res.json();
         console.log(data);
         return data;
-      } catch (error) { }
+      } catch (error) {}
     };
     toast.showAsync(submit, {
       loading: {
