+++ client/src/pages/groups/group/GroupPanel.tsx
@@ -32,7 +32,7 @@ const GroupPanel = () => {
       const data = await res.json();
       console.log(data);
       return data;
-    } catch (error) { }
+    } catch (error) {}
   };
 
   return (
